requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
requests
Pillow
python-dotenv